Create REVERSE INTEGER 7.java
LEETCODE 7 ONE OF THE GOOD APPROACH. 

The main learning that we should know for this problem is any where in your program if you are generating a value that is greater than the integer limits , the java compiler by default shrinks the over flowed value to an integer value with out giving any error or exception. 

For example :

reverse= 214748364 [In Integer limit]

if((reverse*10)+8>Integer.MAX_VALUE)  ----> The result of this arithmetic operation (reverse*10)+8  it self will be stored in an variable with integer data type behind the scene. Hence Java compiler simply shrinks the number to integer value.